Split up nrfUtilToolchainManager
