+++ src/Manager/nrfutil/handleChunk.ts
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) 2015 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
+ */
+
+import type { TaskEvent } from '../../state';
+
+export default (onUpdate: (update: TaskEvent) => void) => {
+    let buffer = '';
+    return (chunk: Buffer) => {
+        buffer += chunk.toString('utf8');
+
+        while (buffer.includes('\n')) {
+            const message = buffer.split('\n')[0];
+            buffer = buffer.substring(message.length + 1);
+            onUpdate(JSON.parse(message));
+        }
+    };
+};
