+++ src/Manager/nrfutil/install.ts
@@ -1,0 +1,24 @@
+/*
+ * Copyright (c) 2015 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
+ */
+
+import { spawn } from 'child_process';
+
+import type { TaskEvent } from '../../state';
+import handleChunk from './handleChunk';
+import nrfutilToolchainManager from './nrfutilToolchainManager';
+
+export default (version: string, onUpdate: (update: TaskEvent) => void) =>
+    new Promise<void>((resolve, reject) => {
+        const tcm = spawn(nrfutilToolchainManager(), [
+            '--json',
+            'install',
+            version,
+        ]);
+
+        tcm.stdout.on('data', handleChunk(onUpdate));
+
+        tcm.on('close', code => (code === 0 ? resolve() : reject()));
+    });
