+++ src/Manager/nrfutil/logVersion.ts
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2015 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
+ */
+
+import { spawnSync } from 'child_process';
+import { logger } from 'pc-nrfconnect-shared';
+
+import nrfutilToolchainManager from './nrfutilToolchainManager';
+
+interface VersionInformation {
+    build_timestamp: string;
+    commit_date: string;
+    commit_hash: string;
+    dependencies: null;
+    host: string;
+    name: string;
+    version: string;
+}
+
+export default () => {
+    const tcm = spawnSync(nrfutilToolchainManager(), ['--json', '--version'], {
+        encoding: 'utf8',
+    });
+
+    const version = JSON.parse(tcm.stdout).data as VersionInformation;
+
+    logger.info(
+        `${version.name} ${version.version} (${version.commit_hash} ${version.commit_date})`
+    );
+};
